Enable privateEditable by default
This is a useful option because it's quite common to forget to make something public, and then when you go to use it instead of autocompleting successfully, giving an error (and potentially an auto-fix in future), and ctrl-clicking the item to jump to it... you have to manually nagivate to the item and add `pub`. It's much more tedious.

Turning it on by default also makes it more discoverable. Because I assumed that everybody would want this behaviour if available I didn't even consider that there might be an option. I only discovered that there was when I tried to implement it myself!

I think one thing that would improve this is if the autocomplete popup said which items were private. Currently there is no indication.