Lower Return Position Impl Trait In Traits correctly
Instead of lowering them as opaque type, in the trait they should get lowered into a synthesized associated type.

Opaques "mostly worked" here, but they break when trying to implement Return Type Notation, which is essentially a way to refer in code to this synthesized associated type. So we need to do the correct thing.