Compute values of RPITITs in impls on a separate query
The previous setup led to a cycle in scenario like the following:
```rust
trait Trait {
    type Assoc;
    fn foo() -> impl Sized;
}
impl Trait for () {
    type Assoc = ();
    fn foo() -> Self::Assoc;
}
```
Because we asked Chalk to normalize the return type of the method, and for that it asked us the datum of the impl, which again causes us to evaluate the RPITITs.

Instead, we only put the associated type ID in `impl_datum()`, and we compute it on `associated_ty_value()`.

This still causes a cycle because Chalk needlessly evaluates all associated types for an impl, but at least it'll work for the new trait solver, and compiler-errors will try to fix that for Chalk too.