Fix new incremental benchmarks
This benchmarks don't measure what they were supposed to, but that's a matter for another time.

The reason we *add* a query to the list is because previously the attrs in the signature queries were computed directly from the item tree and now they call `db.attrs()`. The reason we don't *remove* any query is because, as I said, those benchmarks don't measure what they should