Record inferred const for type or const placeholders
This extends InferenceResult with a new const_of_const_placeholders map
to record which const a given const or type placeholder gets resolved
to.

This is basically the same as type_of_type_placeholder, but for consts.

Note that while surprising, type placeholders TypeRef::Placeholder may
indeed resolve to a const.
In the case of an array, we have an Expr::Underscore as expected:
```
let a: [_; _] = [4, 5, 6];
      //^ TypeRef::Placeholder that resolves to the type i32
         //^ Expr::Underscore that resolves to the const 3
```
But for a struct we have a TypeRef::Placeholder:
```
struct S<T, const N: usize>([T; N]);

let s: S<_, _> = S([1, 2, 3]);
       //^ TypeRef::Placeholder that resolves to the type i32
          //^ TypeRef::Placeholder that resolves to the const 3
```