references.excludeTests to exclude #[cfg(test)]-gated refs
The references.excludeTests setting fails to exclude references
from test code in various scenarios. For example when using
textDocument/references on "foo" in

	pub fn foo() {}
	#[cfg(test)]
	mod tests {
		use crate::foo; // Should be excluded
	}

same when the test module lives in a different file:

	#[cfg(test)]
	mod tests;

Try to fix that for most real-world cases.

TODO: the unit test does not actually test this because whereas
full rust-analyzer finds references in code that's disabled due to
#[cfg(test)] directives, the test logic doesn't; "ModuleData::children"
is always empty. Would appreciate help with that.

changelog fix

Closes #18573