Prepare for compressing spans to save memory
By making token trees span completely opaque outside the `tt` crate: that is, outside it, you cannot get the `span` of a token tree piece anymore, instead you get a `SpannedX` wrapperr that provides a `span()` method (or `open_span()` and `close_span()` for subtrees).

For now these are just markers, but in the future the span will be split and parts of it will be stored compressed in the parent `TopSubtree`. The current change will allow us to do it while only changing the `tt` crate.